# docs/swagger.yaml
basePath: /
definitions:
  main.DeleteSuccess:
    properties:
      status:
        example: deleted
        type: string
    type: object
  main.InvalidBodyError:
    properties:
      error:
        example: invalid body
        type: string
    type: object
  main.NotFoundError:
    properties:
      error:
        example: not found
        type: string
    type: object
  main.User:
    properties:
      age:
        type: integer
      id:
        type: string
      name:
        type: string
    type: object
  main.UserAlreadyExistsError:
    properties:
      error:
        example: user already exists
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: n4mchun@gmail.com
    name: n4mchun
  description: This is a sample swagger for Fiber
  license:
    name: MIT
    url: None
  termsOfService: http://swagger.io/terms/
  title: Fiber Example API
  version: "1.0"
paths:
  /users:
    get:
      consumes:
      - application/json
      description: get all users data
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.User'
            type: array
      summary: Get All Users
      tags:
      - User
    post:
      consumes:
      - application/json
      description: create User by id, name, age
      parameters:
      - description: User Object
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/main.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.User'
        "400":
          description: Invalid Body
          schema:
            $ref: '#/definitions/main.InvalidBodyError'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/main.UserAlreadyExistsError'
      summary: Create User
      tags:
      - User
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: delete user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted
          schema:
            $ref: '#/definitions/main.DeleteSuccess'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/main.NotFoundError'
      summary: Delete User
      tags:
      - User
    get:
      consumes:
      - application/json
      description: get user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.User'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/main.NotFoundError'
      summary: Get User
      tags:
      - User
    put:
      consumes:
      - application/json
      description: update user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User Object
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/main.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.User'
        "400":
          description: Invalid body
          schema:
            $ref: '#/definitions/main.InvalidBodyError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/main.NotFoundError'
      summary: Update User
      tags:
      - User
swagger: "2.0"